fix: replace converter functions to main file
